Distinguish between neighbor id and ip, remove config
The "router-id" in the spine devices is set to the IP that the spine
devices uses to communicate with other spine devices. This router-id is
also the "neighbor" id that a bgp process would use to connect. But the
bgp node's "next-hop" uses the IP of the spine device to which it is
connected to, which is different from the router-id. Thus routes must
use that ip address.

Some config files that were used by setup shell scripts are being
removed as they're no longer used